+++ CHANGELOG.md
@@ -1,13 +1,34 @@
+## [2.0.0]
+
+- BREAKING: SuperclusterImmutableLayer is now SuperclusterLayer.immutable.
+- BREAKING: SuperclusterMutableLayer is now SuperclusterLayer.mutable.
+- BREAKING: SuperclusterController's aggregatedClusterDataStream has been
+            replaced with stateStream which contains a new SuperclusterState
+            object containing the aggregated cluster data and the loading state.
+- BREAKING: SuperclusterController's `all()` method now returns a
+            `Future<Iterable<Marker>>` instead of an `Iterable<Marker>` as it
+            is possible that the index is still loading.
+- BREAKING: Creation of the supercluster index is now done in a separate isolate
+  by default using Flutter's `compute` function. This can be changed with the
+  new `wrapIndexCreation` option. Note due to how `compute` works the Marker
+  instances inside the index will be clones of the original Markers and will
+  not be equal to or have the same hashCode as the original Markers. If you need
+  to show popups or perform other actions which required matching against the
+  original Marker instances you should implement an equals and hashcode or use
+  a library like equatable which simplifies doing so.
+- Whilst the index is loading a loading overlay is displayed. This overlay can
+  be customised with the new `loadingOverlayBuilder` option.
+
 ## [1.0.0]
 
-- BREAKING: Now requires flutter_map 3.0.0 or higher.
-- BREAKING: Rename the package to flutter_map_supercluster:
-    - FlutterMapFastCluster becomes SuperclusterImmutableLayer
+- BREAKING: Now requires `flutter_map` 3.0.0 or higher.
+- BREAKING: Renamed the package to `flutter_map_supercluster`:
+    - `FlutterMapFastCluster` becomes `SuperclusterImmutableLayer`
     - etc. (see example if you are unsure).
-- Added SuperclusterMutable which allows adding/removing markers efficiently. It
-  is no longer necessary to rebuild the whole index when changing markers.
+- Added `SuperclusterMutable` which allows adding/removing markers efficiently.
+  It is no longer necessary to rebuild the whole index when changing markers.
 - Changing the markers is now done via the respective controller
-  (SuperclusterImmutableController/SuperclusterMutableController).
+  (`SuperclusterImmutableController`/`SuperclusterMutableController`).
 
 ## [0.0.2]